--- a/field_cache.go
+++ b/field_cache.go
@@ -2,8 +2,10 @@ package validator
 
 type (
 	fieldCache struct {
+		index     int
 		isPrivate bool
 		name      string
 		tagValue  string
+		tagChunk  *tagChunk
 	}
 )